requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Markdown==3.7
pillow==11.1.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
